[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cc
@@ -5,7 +5,7 @@
 #include <array>
 using namespace std;
 
-int calculate(int A, int B) {
+long long calculate(long long A, long long B) {
     return (A + B) * (A - B);
 }
 
@@ -14,7 +14,7 @@ int main(void) {
     ios_base::sync_with_stdio(false);
     
 
-    int A, B;
+    long long A, B;
     cin >> A >> B;
     cout << calculate(A, B);
